[update] models

+++ lib/layouts/pages/profile-pages/profile-overview/profile_overview_sec3.dart
@@ -42,10 +42,10 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
               child: Container(
                 color: ColorPicker.kGreyLight3,
                 width: Responsive.isDesktop(context)
-                    ? SS.sW(context) * .50 as double
+                    ? SS.sW(context) * .50
                     : Responsive.isMobile(context)
-                        ? SS.sW(context) * .90 as double
-                        : SS.sW(context) * .60 as double,
+                        ? SS.sW(context) * .90
+                        : SS.sW(context) * .60,
                 child: SingleChildScrollView(
                   child: Column(
                     children: [
@@ -103,17 +103,17 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
               child: Container(
                 color: ColorPicker.kGreyLight8,
                 width: Responsive.isDesktop(context)
-                    ? SS.sW(context) * .50 as double
+                    ? SS.sW(context) * .50
                     : Responsive.isMobile(context)
-                        ? SS.sW(context) * .90 as double
-                        : SS.sW(context) * .60 as double,
+                        ? SS.sW(context) * .90
+                        : SS.sW(context) * .60,
                 // height: 375,
                 child: SingleChildScrollView(
                   child: Column(
                     children: [
                       PopUpHeadMenu.popUpHead('Expertise', context),
                       SizedBox(
-                        height: SS.sH(context) * .05 as double,
+                        height: SS.sH(context) * .05,
                       ),
                       Padding(
                         padding: EdgeInsets.all(16.0),
@@ -122,10 +122,10 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
                             Container(
                               color: ColorPicker.kWhite,
                               width: Responsive.isDesktop(context)
-                                  ? SS.sW(context) * .50 as double
+                                  ? SS.sW(context) * .50
                                   : Responsive.isMobile(context)
-                                      ? SS.sW(context) * .90 as double
-                                      : SS.sW(context) * .60 as double,
+                                      ? SS.sW(context) * .90
+                                      : SS.sW(context) * .60,
                               child: Padding(
                                 padding: EdgeInsets.all(20),
                                 child: Column(
@@ -208,8 +208,7 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
                                                         color: ColorPicker
                                                             .kBlueLight3)),
                                                 alignment: Alignment.center,
-                                                width: SS.sW(context) * .091
-                                                    as double,
+                                                width: SS.sW(context) * .091,
                                                 height: 36,
                                                 child: Text(
                                                   'Add to list',
@@ -330,10 +329,10 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
               child: Container(
                 color: ColorPicker.kGreyLight3,
                 width: Responsive.isDesktop(context)
-                    ? SS.sW(context) * .50 as double
+                    ? SS.sW(context) * .50
                     : Responsive.isMobile(context)
-                        ? SS.sW(context) * .90 as double
-                        : SS.sW(context) * .60 as double,
+                        ? SS.sW(context) * .90
+                        : SS.sW(context) * .60,
                 child: SingleChildScrollView(
                   child: Column(
                     mainAxisAlignment: MainAxisAlignment.start,
@@ -343,10 +342,10 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
                         padding: EdgeInsets.all(16.0),
                         child: Container(
                           width: Responsive.isDesktop(context)
-                              ? SS.sW(context) * .50 as double
+                              ? SS.sW(context) * .50
                               : Responsive.isMobile(context)
-                                  ? SS.sW(context) * .90 as double
-                                  : SS.sW(context) * .60 as double,
+                                  ? SS.sW(context) * .90
+                                  : SS.sW(context) * .60,
                           color: ColorPicker.kWhite,
                           child: Row(
                             crossAxisAlignment: CrossAxisAlignment.start,
@@ -641,7 +640,7 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
                     onPressed: () {},
                     child: Padding(
                       padding: EdgeInsets.symmetric(
-                        vertical: SS.sH(context) * 0.010 as double,
+                        vertical: SS.sH(context) * 0.010,
                       ),
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.center,
@@ -652,7 +651,7 @@ class _ProfileOverviewSec3State extends State<ProfileOverviewSec3> {
                             size: 12,
                           ),
                           SizedBox(
-                            width: SS.sW(context) * .01 as double,
+                            width: SS.sW(context) * .01,
                           ),
                           AutoSizeText(
                             'Grow your network',
